+++ crates/papyrus_gateway/src/api/mod.rs
@@ -15,9 +15,9 @@ use starknet_api::transaction::EventKey;
 
 use self::v0_3_0::v0_3_0_impl::JsonRpcServerV0_3_0Impl;
 use self::v0_3_0::JsonRpcV0_3_0Server;
-use crate::transaction::Event;
 use crate::v0_3_0::deprecated_contract_class::ContractClass as DeprecatedContractClass;
 use crate::v0_3_0::state::ContractClass;
+use crate::v0_3_0::transaction::Event;
 
 #[derive(Copy, Clone, Debug, Deserialize, Eq, PartialEq, Serialize)]
 pub enum Tag {
