+++ crates/papyrus_gateway/src/api/v0_3_0/v0_3_0_test.rs
@@ -35,13 +35,13 @@ use crate::api::{
 use crate::test_utils::{
     get_starknet_spec_api_schema, get_test_gateway_config, get_test_rpc_server_and_storage_writer,
 };
-use crate::transaction::{
-    Event, TransactionOutput, TransactionReceipt, TransactionReceiptWithStatus, TransactionStatus,
-    TransactionWithType, Transactions,
-};
 use crate::v0_3_0::block::Block;
 use crate::v0_3_0::deprecated_contract_class::ContractClass as DeprecatedContractClass;
 use crate::v0_3_0::state::{ContractClass, StateUpdate, ThinStateDiff};
+use crate::v0_3_0::transaction::{
+    Event, TransactionOutput, TransactionReceipt, TransactionReceiptWithStatus, TransactionStatus,
+    TransactionWithType, Transactions,
+};
 use crate::{run_server, ContinuationTokenAsStruct};
 
 #[tokio::test]
