+++ crates/papyrus_gateway/src/gateway_test.rs
@@ -17,10 +17,6 @@ use starknet_api::block::{BlockHash, BlockHeader, BlockNumber, BlockStatus};
 use starknet_api::deprecated_contract_class::{
     EventAbiEntry, FunctionAbiEntryWithType, StructAbiEntry,
 };
-use starknet_api::transaction::{
-    DeclareTransactionOutput, DeployAccountTransactionOutput, DeployTransactionOutput,
-    InvokeTransactionOutput, L1HandlerTransactionOutput, Transaction,
-};
 use test_utils::{get_rng, GetTestInstance};
 use tower::BoxError;
 
@@ -28,7 +24,6 @@ use crate::api::version_config::{LATEST_VERSION_ID, VERSION_CONFIG};
 use crate::api::JsonRpcError;
 use crate::middleware::proxy_request;
 use crate::test_utils::get_test_gateway_config;
-use crate::transaction::{TransactionOutput, TransactionReceipt};
 use crate::v0_3_0::deprecated_contract_class::{
     ContractClassAbiEntryType, ContractClassAbiEntryWithType,
 };
@@ -213,32 +208,3 @@ macro_rules! test_contract_class_abi_entry_with_type {
 test_contract_class_abi_entry_with_type!(Event, EventAbiEntry);
 test_contract_class_abi_entry_with_type!(Function, FunctionAbiEntryWithType);
 test_contract_class_abi_entry_with_type!(Struct, StructAbiEntry);
-
-macro_rules! test_recipe_from_transtaction_output {
-    ($variant:ident, $recipe_type:ident) => {
-        paste! {
-            #[tokio::test]
-            async fn [<test_recipe_from_transtaction_output_ $variant:lower>]() {
-                let mut rng = get_rng();
-                let block_header = BlockHeader::default();
-                let transaction = Transaction::$variant(
-                    starknet_api::transaction::[<$variant Transaction>]::get_test_instance(&mut rng),
-                );
-                let output = TransactionOutput::$variant([<$variant TransactionOutput>]::default());
-                let receipt = TransactionReceipt::from_transaction_output(
-                    output,
-                    &transaction,
-                    block_header.block_hash,
-                    block_header.block_number,
-                );
-                assert_matches!(receipt, TransactionReceipt::$recipe_type(_));
-            }
-        }
-    }
-}
-
-test_recipe_from_transtaction_output!(Declare, Common);
-test_recipe_from_transtaction_output!(Invoke, Common);
-test_recipe_from_transtaction_output!(L1Handler, Common);
-test_recipe_from_transtaction_output!(Deploy, Deploy);
-test_recipe_from_transtaction_output!(DeployAccount, Deploy);
