+++ crates/papyrus_gateway/src/v0_3_0/block.rs
@@ -2,7 +2,7 @@ use serde::{Deserialize, Serialize};
 use starknet_api::block::{BlockHash, BlockNumber, BlockStatus, BlockTimestamp};
 use starknet_api::core::{ContractAddress, GlobalRoot};
 
-use crate::transaction::Transactions;
+use crate::v0_3_0::transaction::Transactions;
 
 #[derive(Debug, Clone, Eq, PartialEq, Hash, Deserialize, Serialize, PartialOrd, Ord)]
 pub struct BlockHeader {
