+++ crates/papyrus_gateway/src/v0_3_0/mod.rs
@@ -1,3 +1,6 @@
 pub mod block;
 pub mod deprecated_contract_class;
 pub mod state;
+pub mod transaction;
+#[cfg(test)]
+mod transaction_test;
